requirements.txt
Django
djangorestframework
gunicorn
whitenoise
Pillow
psycopg2-binary
dj-database-url
